fixed mp_uint_t type error

+++ ports/k210-standalone/mpconfigport.h
@@ -211,8 +211,8 @@ extern const struct _mp_print_t mp_debug_print;
 // to print such value. So, we avoid int32_t and use int directly.
 #define UINT_FMT "%u"
 #define INT_FMT "%d"
-typedef int mp_int_t; // must be pointer size
-typedef unsigned mp_uint_t; // must be pointer size
+typedef int64_t mp_int_t; // must be pointer size
+typedef uint64_t mp_uint_t; // must be pointer size
 
 typedef long mp_off_t;
